Saved Variables have defeated me
This has code that theoretically should work to default the dropdowns to the last selected item, but I can't get it to work, itll populate the savedvariables file, you can even use a script in game to reference the variables and they are recognized. I'm assuming there's something wrong with the dynamic aspect of oogleCraftDropdownFrameName, but who knows.